requirements.txt
python-dotenv
requests
json
RPi.GPIO